move loadURL tests
