+++ spec-main/api-browser-window-spec.ts
@@ -2,9 +2,10 @@ import * as chai from 'chai'
 import * as chaiAsPromised from 'chai-as-promised'
 import * as path from 'path'
 import * as fs from 'fs'
+import * as qs from 'querystring'
 import * as http from 'http'
 import { AddressInfo } from 'net'
-import { app, BrowserWindow } from 'electron'
+import { app, BrowserWindow, ipcMain, OnBeforeSendHeadersListenerDetails } from 'electron'
 import { emittedOnce } from './events-helpers';
 import { closeWindow } from './window-helpers';
 
@@ -173,6 +174,200 @@ describe('BrowserWindow module', () => {
       windows.forEach(win => win.focus())
       windows.forEach(win => win.destroy())
       app.removeListener('browser-window-focus', focusListener)
+    })
+  })
+
+  describe('BrowserWindow.loadURL(url)', () => {
+    let w = null as unknown as BrowserWindow
+    beforeEach(() => {
+      w = new BrowserWindow({show: false, webPreferences: {nodeIntegration: true}})
+    })
+    afterEach(async () => {
+      await closeWindow(w)
+      w = null as unknown as BrowserWindow
+    })
+    let server = null as unknown as http.Server
+    let url = null as unknown as string
+    let postData = null as any
+    before((done) => {
+      const filePath = path.join(fixtures, 'pages', 'a.html')
+      const fileStats = fs.statSync(filePath)
+      postData = [
+        {
+          type: 'rawData',
+          bytes: Buffer.from('username=test&file=')
+        },
+        {
+          type: 'file',
+          filePath: filePath,
+          offset: 0,
+          length: fileStats.size,
+          modificationTime: fileStats.mtime.getTime() / 1000
+        }
+      ]
+      server = http.createServer((req, res) => {
+        function respond () {
+          if (req.method === 'POST') {
+            let body = ''
+            req.on('data', (data) => {
+              if (data) body += data
+            })
+            req.on('end', () => {
+              const parsedData = qs.parse(body)
+              fs.readFile(filePath, (err, data) => {
+                if (err) return
+                if (parsedData.username === 'test' &&
+                    parsedData.file === data.toString()) {
+                  res.end()
+                }
+              })
+            })
+          } else if (req.url === '/302') {
+            res.setHeader('Location', '/200')
+            res.statusCode = 302
+            res.end()
+          } else if (req.url === '/navigate-302') {
+            res.end(`<html><body><script>window.location='${url}/302'</script></body></html>`)
+          } else if (req.url === '/cross-site') {
+            res.end(`<html><body><h1>${req.url}</h1></body></html>`)
+          } else {
+            res.end()
+          }
+        }
+        setTimeout(respond, req.url && req.url.includes('slow') ? 200 : 0)
+      })
+      server.listen(0, '127.0.0.1', () => {
+        url = `http://127.0.0.1:${(server.address() as AddressInfo).port}`
+        done()
+      })
+    })
+
+    after(() => {
+      server.close()
+    })
+
+    it('should emit did-start-loading event', (done) => {
+      w.webContents.on('did-start-loading', () => { done() })
+      w.loadURL('about:blank')
+    })
+    it('should emit ready-to-show event', (done) => {
+      w.on('ready-to-show', () => { done() })
+      w.loadURL('about:blank')
+    })
+    it('should emit did-fail-load event for files that do not exist', (done) => {
+      w.webContents.on('did-fail-load', (event, code, desc, url, isMainFrame) => {
+        expect(code).to.equal(-6)
+        expect(desc).to.equal('ERR_FILE_NOT_FOUND')
+        expect(isMainFrame).to.equal(true)
+        done()
+      })
+      w.loadURL('file://a.txt')
+    })
+    it('should emit did-fail-load event for invalid URL', (done) => {
+      w.webContents.on('did-fail-load', (event, code, desc, url, isMainFrame) => {
+        expect(desc).to.equal('ERR_INVALID_URL')
+        expect(code).to.equal(-300)
+        expect(isMainFrame).to.equal(true)
+        done()
+      })
+      w.loadURL('http://example:port')
+    })
+    it('should set `mainFrame = false` on did-fail-load events in iframes', (done) => {
+      w.webContents.on('did-fail-load', (event, code, desc, url, isMainFrame) => {
+        expect(isMainFrame).to.equal(false)
+        done()
+      })
+      w.loadFile(path.join(fixtures, 'api', 'did-fail-load-iframe.html'))
+    })
+    it('does not crash in did-fail-provisional-load handler', (done) => {
+      w.webContents.once('did-fail-provisional-load', () => {
+        w.loadURL('http://127.0.0.1:11111')
+        done()
+      })
+      w.loadURL('http://127.0.0.1:11111')
+    })
+    it('should emit did-fail-load event for URL exceeding character limit', (done) => {
+      w.webContents.on('did-fail-load', (event, code, desc, url, isMainFrame) => {
+        expect(desc).to.equal('ERR_INVALID_URL')
+        expect(code).to.equal(-300)
+        expect(isMainFrame).to.equal(true)
+        done()
+      })
+      const data = Buffer.alloc(2 * 1024 * 1024).toString('base64')
+      w.loadURL(`data:image/png;base64,${data}`)
+    })
+
+    it('should return a promise', () => {
+      const p = w.loadURL('about:blank')
+      expect(p).to.have.property('then')
+    })
+
+    it('should return a promise that resolves', async () => {
+      expect(w.loadURL('about:blank')).to.eventually.be.fulfilled
+    })
+
+    it('should return a promise that rejects on a load failure', async () => {
+      const data = Buffer.alloc(2 * 1024 * 1024).toString('base64')
+      const p = w.loadURL(`data:image/png;base64,${data}`)
+      await expect(p).to.eventually.be.rejected
+    })
+
+    it('should return a promise that resolves even if pushState occurs during navigation', async () => {
+      const p = w.loadURL('data:text/html,<script>window.history.pushState({}, "/foo")</script>')
+      await expect(p).to.eventually.be.fulfilled
+    })
+
+    describe('POST navigations', () => {
+      afterEach(() => { w.webContents.session.webRequest.onBeforeSendHeaders(null) })
+
+      it('supports specifying POST data', async () => {
+        await w.loadURL(url, { postData: postData })
+      })
+      it('sets the content type header on URL encoded forms', async () => {
+        await w.loadURL(url)
+        const requestDetails: Promise<OnBeforeSendHeadersListenerDetails> = new Promise(resolve => {
+          w.webContents.session.webRequest.onBeforeSendHeaders((details, callback) => {
+            resolve(details)
+          })
+        })
+        w.webContents.executeJavaScript(`
+          form = document.createElement('form')
+          document.body.appendChild(form)
+          form.method = 'POST'
+          form.submit()
+        `)
+        const details = await requestDetails
+        expect(details.requestHeaders['Content-Type']).to.equal('application/x-www-form-urlencoded')
+      })
+      it('sets the content type header on multi part forms', async () => {
+        await w.loadURL(url)
+        const requestDetails: Promise<OnBeforeSendHeadersListenerDetails> = new Promise(resolve => {
+          w.webContents.session.webRequest.onBeforeSendHeaders((details, callback) => {
+            resolve(details)
+          })
+        })
+        w.webContents.executeJavaScript(`
+          form = document.createElement('form')
+          document.body.appendChild(form)
+          form.method = 'POST'
+          form.enctype = 'multipart/form-data'
+          file = document.createElement('input')
+          file.type = 'file'
+          file.name = 'file'
+          form.appendChild(file)
+          form.submit()
+        `)
+        const details = await requestDetails
+        expect(details.requestHeaders['Content-Type'].startsWith('multipart/form-data; boundary=----WebKitFormBoundary')).to.equal(true)
+      })
+    })
+
+    it('should support support base url for data urls', (done) => {
+      ipcMain.once('answer', (event, test) => {
+        expect(test).to.equal('test')
+        done()
+      })
+      w.loadURL('data:text/html,<script src="loaded-from-dataurl.js"></script>', { baseURLForDataURL: `file://${path.join(fixtures, 'api')}${path.sep}` })
     })
   })
 })
